requirements.txt
gymnasium>=0.29.0
numpy>=1.26.0
torch>=2.1.0
sqlparse>=0.4.4
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
tqdm>=4.66.0
rich>=13.7.0
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
jupyter>=1.0.0
ipykernel>=6.0.0
